+++ tests/WasmSharp.Tests/Modules/ModuleDecoder_DecodeFailureTests.cs
@@ -34,16 +34,12 @@ internal partial class ModuleDecoder_DecodeTests
                 .That(exception.UnverifiedRanges[0].Stage)
                 .IsEqualTo(WasmProcessingStage.Decode);
             await Assert.That(exception.UnverifiedRanges[0].StartOffset).IsEqualTo(8L);
-            await Assert
-                .That(exception.UnverifiedRanges[0].EndOffset)
-                .IsEqualTo((long)bytes.Length);
+            await Assert.That(exception.UnverifiedRanges[0].EndOffset).IsEqualTo(bytes.Length);
             await Assert
                 .That(exception.UnverifiedRanges[1].Stage)
                 .IsEqualTo(WasmProcessingStage.Validate);
             await Assert.That(exception.UnverifiedRanges[1].StartOffset).IsEqualTo(0L);
-            await Assert
-                .That(exception.UnverifiedRanges[1].EndOffset)
-                .IsEqualTo((long)bytes.Length);
+            await Assert.That(exception.UnverifiedRanges[1].EndOffset).IsEqualTo(bytes.Length);
         }
     }
 
@@ -81,16 +77,12 @@ internal partial class ModuleDecoder_DecodeTests
                 .That(exception.UnverifiedRanges[0].Stage)
                 .IsEqualTo(WasmProcessingStage.Decode);
             await Assert.That(exception.UnverifiedRanges[0].StartOffset).IsEqualTo(33L);
-            await Assert
-                .That(exception.UnverifiedRanges[0].EndOffset)
-                .IsEqualTo((long)bytes.Length);
+            await Assert.That(exception.UnverifiedRanges[0].EndOffset).IsEqualTo(bytes.Length);
             await Assert
                 .That(exception.UnverifiedRanges[1].Stage)
                 .IsEqualTo(WasmProcessingStage.Validate);
             await Assert.That(exception.UnverifiedRanges[1].StartOffset).IsEqualTo(0L);
-            await Assert
-                .That(exception.UnverifiedRanges[1].EndOffset)
-                .IsEqualTo((long)bytes.Length);
+            await Assert.That(exception.UnverifiedRanges[1].EndOffset).IsEqualTo(bytes.Length);
         }
     }
